requirements.txt
numpy
pytest==5.4.1
dill
python-mnist
flaky